+++ src/LabelPrint.Application/Validation/PrinterUpsertDtoValidator.cs
@@ -14,6 +14,8 @@ public sealed class PrinterUpsertDtoValidator : AbstractValidator<PrinterUpsertD
         RuleFor(x => x.Name).NotEmpty().MaximumLength(128);
         RuleFor(x => x.ConnectionString).MaximumLength(512);
         RuleFor(x => x.PaperWidthMm).InclusiveBetween(10, 300);
+        RuleFor(x => x.PrintOffsetXMm).InclusiveBetween(-20, 20);
+        RuleFor(x => x.PrintOffsetYMm).InclusiveBetween(-20, 20);
         RuleFor(x => x.Dpi).InclusiveBetween(72, 600);
         RuleFor(x => x.Darkness).InclusiveBetween(0, 15);
         RuleFor(x => x.Speed).InclusiveBetween(1, 10);